refactor: extract env files logic into EnvGenerator
